Create module Groupable
