+++ lib/stat_tracker.rb
@@ -1,8 +1,10 @@
 require_relative "./game_teams_stats"
 require_relative "./game_stats"
 require_relative "./teams_stats"
+require './groupable'
 
 class StatTracker
+  include Groupable
   attr_reader :game_stats,
               :teams_stats,
               :game_teams_stats
